feat: refactor product prices and register new sale with get sales

+++ src/main/java/org/sistema/arroz/riceservice/modules/products/application/port/in/ProductToRegister.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class ProductToRegister {
     String productName;
     Double sacks;
-    Double stock;
+    Integer stock;
     Long communityId;
 }
